10/4/2022

+++ src/app/admin/pages/account/account.component.ts
@@ -1,6 +1,6 @@
 import { AccountService } from './../../../services/account.service';
 import { Component, OnInit } from '@angular/core';
-
+import Swal from 'sweetalert2'
 @Component({
   selector: 'app-account',
   templateUrl: './account.component.html',
@@ -11,6 +11,7 @@ export class AccountComponent implements OnInit {
   allAccount:any;
   constructor(private accService: AccountService) { }
   ngOnInit(): void {
+    document.documentElement.scrollTop = 0;
     this.getAllAccount();
   }
   getAllAccount(){
@@ -25,8 +26,26 @@ export class AccountComponent implements OnInit {
     this.accService.editItem(id,data).subscribe();
   }
   delete(id:number){
-    this.accService.deleteItem(id).subscribe(()=>{
-      this.getAllAccount();
-    });
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+        this.accService.deleteItem(id).subscribe(()=>{
+          this.getAllAccount();
+        });
+      }
+    })
+    
   }
 }
